starship: Ignore `.python-version` files
Don't assume the existence of a `.python-version` file indicates that
the directory is a Python project. It is useful to have a
`.python-version` that trickles down the directory tree without the
directory itself being a python project (for example, to set an
organization-wide python version).

This probably applies to other `-version` files as well.