stow: Always target the user's home directory
Previously this repository would have to be located as an immediate
subdirectory of the user's `$HOME` in order for Stow to place dotfiles
in the correct location.

We now use the `.stowrc` file to specify that the target directory is
the `$HOME` directory.

Care must be taken to run Stow as the intended user with the intended
`$HOME`.