zsh: Use Git submodules for plugins
Before this patch, I was using system packages for the zsh plugins
`zsh-autosuggestions` and `zsh-syntax-highlighting`. Now, they're Git
submodules. This makes the shell behaves more or less the same
regardless of which version is available (if any) in the system's
package manager. The downside is that they have to be updated manually.