gh+1password+shell: Add 1Password CLI support
This patch adds support for 1Password CLI plugins. It supports biometric
authentication for a number of CLI tools like `gh`.

We used to have a `gh` alias that would make sure that the `GH_TOKEN`
variable was unset, forcing `gh` to use its built-in authentication.
That was a bit of a hack and didn't work with the 1Password plugin, so
it's been removed.